+++ packages/message-encryption/package.json
@@ -16,6 +16,10 @@
     "./symmetric": {
       "types": "./dist/symmetric.d.ts",
       "import": "./dist/symmetric.js"
+    },
+    "./crypto": {
+      "types": "./dist/crypto/index.d.ts",
+      "import": "./dist/crypto/index.js"
     }
   },
   "typesVersions": {
