+++ packages/message-encryption/src/misc.ts
@@ -8,3 +8,12 @@ export const Symmetric = {
 export const Asymmetric = {
   keySize: 32
 };
+
+export const OneMillion = BigInt(1_000_000);
+
+export const Version = 1;
+
+export type Signature = {
+  signature: Uint8Array;
+  publicKey: Uint8Array | undefined;
+};
